requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
filelock==3.18.0
fsspec==2025.5.1
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python-headless==4.11.0.86
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
PyYAML==6.0.2
scipy==1.16.0
setuptools==80.9.0
simsimd==6.4.9
stringzilla==3.12.5
sympy==1.14.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0